backend/requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
flask-cors==3.0.10
flask-socketio==5.1.1
psycopg2-binary==2.9.1
python-dotenv==0.19.0
PyJWT==2.1.0
bcrypt==3.2.0
gunicorn==20.1.0
python-socketio==5.4.0
eventlet==0.33.0
dnspython==2.2.1
pyzmq==26.3.0 